+++ src/utils.h
@@ -3,15 +3,17 @@
 
 #include <stdbool.h>
 
+
+// Two numbers with extra steps
 typedef struct {
-  float min;
-  float max;
+  double min;
+  double max;
 } Interval;
 
-Interval interval_new(float min, float max);
-bool interval_surrounds(Interval interval, float x);
-float interval_clamp(Interval interval, float threshold);
-float interval_sizeof(Interval interval);
+Interval interval_new(double min, double max);
+bool interval_surrounds(Interval interval, double x);
+double interval_clamp(Interval interval, double threshold);
+double interval_sizeof(Interval interval);
 
 #endif
